\name{tbdiag-package}
\alias{tbdiag-package}
\alias{tbdiag}
\docType{package}
\title{
tbdiag: utilities for tuberculosis diagnostics research
}
\description{
This package provides functions to assist researchers working in the field of tuberculosis diagnostics.  Functions for the interpretation of two popular interferon-gamma release assays are provided, and additional functionality is planned.
}
\details{
\tabular{ll}{
Package: \tab tbdiag\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2012-01-06\cr
License: \tab MIT\cr
LazyLoad: \tab yes\cr
}
}
\author{
Matt Parker

Maintainer: Matt Parker <mattmparker@gmail.com>
}

\keyword{ package }
