% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tbdiag-package.r
\docType{data}
\name{test.qfts}
\alias{test.qfts}
\title{A sample of 6,833 real QFT results.}
\description{
A sample of 6,833 real QFT results.
}
\author{
Matt Parker \email{mattmparker@gmail.com}
}
\keyword{data}
\keyword{qft}
\keyword{quantiferon}

