% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tbdiag-package.r
\docType{data}
\name{test.tspots}
\alias{test.tspots}
\title{A sample of 6,434 real TSPOT results.}
\description{
A sample of 6,434 real TSPOT results.
}
\author{
Matt Parker \email{mattmparker@gmail.com}
}
\keyword{data}
\keyword{tspot}

