\name{tbdiag-package}
\alias{tbdiag-package}
\alias{tbdiag}
\docType{package}
\title{
tbdiag: utilities for tuberculosis diagnostics research
}
\description{
This package provides utilities for researchers working in the field of tuberculosis diagnostics.  At the moment, the utilities are functions for the interpretation of the commercial inteferon-gamma release assays the Cellestis Quantiferon and the Oxford Immunotec TSPOT.TB.  Additional suggestions welcome.
}
\details{
\tabular{ll}{
Package: \tab tbdiag\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-05-26\cr
License: \tab MIT\cr
LazyLoad: \tab yes\cr
}
}
\author{
Matt Parker

Maintainer: Matt Parker <mattmparker@gmail.com>
}

\keyword{ package }
