\docType{package}
\name{tbdiag-package}
\alias{tbdiag}
\alias{tbdiag-package}
\title{tbdiag: utilities for tuberculosis diagnostics research}
\description{
  This package provides functions to assist researchers
  working in the field of tuberculosis diagnostics.
  Functions for the interpretation of two popular
  interferon-gamma release assays are provided, and
  additional functionality is planned.
}
\author{
  Matt Parker \email{mattmparker@gmail.com}

  Matt Parker \email{mattmparker@gmail.com}

  Matt Parker \email{mattmparker@gmail.com}
}
\keyword{6,434}
\keyword{6,833}
\keyword{A}
\keyword{QFT}
\keyword{TSPOT}
\keyword{data}
\keyword{of}
\keyword{package}
\keyword{qft}
\keyword{quantiferon}
\keyword{real}
\keyword{results.}
\keyword{sample}
\keyword{tspot}
\keyword{tuberculosis}

