% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tbdiag-package.r
\docType{package}
\name{tbdiag-package}
\alias{tbdiag}
\alias{tbdiag-package}
\title{tbdiag: utilities for tuberculosis diagnostics research}
\description{
This package provides functions to assist researchers working in the field of tuberculosis diagnostics.  Functions for the interpretation of two popular interferon-gamma release assays are provided, and additional functionality is planned.
}
\author{
Matt Parker \email{mattmparker@gmail.com}
}
\keyword{package}
\keyword{tuberculosis}

